reorganizing sample_data
