+++ app/stac_api/sample_data/swissMapRaster200/collection.json
