+++ app/stac_api/sample_data/swissMapRaster200/items/smr200-200-1-2019.json
