+++ app/stac_api/sample_data/swissMapRaster200/items/smr200-200-4-2019.json
