+++ app/stac_api/sample_data/swissTLM3D/collection.json
