+++ app/stac_api/sample_data/swissTLM3D/items/swisstlmregio-2020.json
